Thank you. Your svn details should be with you by the end of day.
Please note that with the svn login - which is the same as your wordpress.org - capitalization of the username/password matters.

If you're new to WordPress and SVN, please read http://make.wordpress.org/plugins/2012/06/09/the-plugins-directory-and-readme-txt-files/ to understand how your readme impacts what shows on your WordPress.org page.

(In order to speed up the process for everyone, there's no need for you to reply to this email. Even if you're just thinking of saying 'thanks' it makes more emails for us to process. We appreciate you understanding!)


git-svn-id: https://plugins.svn.wordpress.org/free-quotation/trunk@816497 b8457f37-d9ea-0310-8a92-e5e31aec5664